Added if Statement for hostapd.conf restore
Added if Statement for restore

+++ mitmAP_rpi.py
@@ -326,7 +326,8 @@ try:
         os.system("sudo screen -S mitmap-driftnet -X stuff '^C\n'")
     if tshark_if == "y" or tshark_if == "":
         os.system("sudo screen -S mitmap-tshark -X stuff '^C\n'")
-    print("[I] Restoring old hostapd...")
+    if os.path.isfile("/etc/hostapd/hostapd.conf.backup"):
+        print("[I] Restoring old hostapd...")
         os.system("sudo mv  /etc/hostapd/hostapd.conf.backup /etc/hostapd/hostapd.conf")
     # print("[I] Restoring old NetworkManager.cfg")
     # os.system("sudo mv /etc/NetworkManager/NetworkManager.conf.backup /etc/NetworkManager/NetworkManager.conf")
@@ -394,7 +395,8 @@ except KeyboardInterrupt:
         os.system("sudo wondershaper clear " + ap_iface + " > /dev/null 2>&1")
     except:
         pass
-    print("[I] Restoring old hostapd...")
+    if os.path.isfile("/etc/hostapd/hostapd.conf.backup"):
+        print("[I] Restoring old hostapd...")
         os.system("sudo mv  /etc/hostapd/hostapd.conf.backup /etc/hostapd/hostapd.conf")
     print("[I] Flushing iptables rules...")
     os.system("sudo iptables --flush")
